requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.2
aiohttp>=3.9.0  # Async HTTP client for Ollama

# Development
pytest>=7.4.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
pylint>=2.17.0
